# Conflicts:
#	CUDA_Noise.vcxproj

+++ cpp/modules/generators/RidgedMulti.h
@@ -1,7 +1,6 @@
 #pragma once
 #ifndef RIDGED_MULTI_H
 #define RIDGED_MULTI_H
-#include "common\CUDA_Include.h"
 #include "../Base.h"
 
 namespace noise {
@@ -27,7 +26,7 @@ namespace noise {
 				int octaves = DEFAULT_RIDGED_OCTAVES, float persist = DEFAULT_RIDGED_PERSISTENCE);
 
 			// Get source module count: must be 0, this is a generator and can't have preceding modules.
-			virtual int GetSourceModuleCount() const override;
+			virtual size_t GetSourceModuleCount() const override;
 
 			// Launches the kernel and fills this object's surface object with the relevant data.
 			virtual void Generate() override;
